Fixed FrcCANTalon so getPosition and getVelocity will always return encoder values
Some programs change the selected sensor of the motor controller, so calling get position during that time will get a weird value. If the feedback device is a quadrature encoder, it will always call the encoder, not the selected feedback device. If it is not a quad encoder, the talon will return the selected device value.