All PDP channels now registered under a default name if not otherwise named
At the end of robotInit(), a new method on FrcPdp is called which assigns
the name "Unassigned_<i>" (where i is the 0-indexed channel name) for
all channels that do not already have an energy usage name set. This
allows easy tracking of energy usage for all PDP channels, named or
not.