Fixed bugs, added point interpolation, added more setters
There was a bug where the no-intersect method of point selection was flipping the inequality.

Also, the intersect method of point selection couldn't select the same point twice.

Also, threshold MUST BE less than followingDistance.

Also, linearly interpolate between points based on the distance between the intersection and the points.