Fixed bug with non-encoder pid sources, also non-holonomic robots
Don't set the reference pose every update, because then the position would always be 0,0 so it would run forever.

Don't do any heading transforms, because then that requires encoders as pid input, while that's not guaranteed.

As it is, this means that heading cannot be controlled independently of x and y, even though holonomic drivebases have that capability. This is due to complications arising from the fact that the pid sources (and the reference frames of those sources) are completely unknown, and to preserve continuity must be assumed to be in the robot local reference frame. With this fix, TrcPidDrive is *almost* exactly like before the new odometry features, as long as heading is never controlled simultaneously with position. If it is, robots will misbehave, potentially running indefinitely.

Since we never did this to begin with, it shouldn't be much of a problem.